requirements.txt
asgiref==3.3.1
Django==3.2.14
django-filter==2.4.0
djangorestframework==3.12.2
djangorestframework-simplejwt==4.6.0
Markdown==3.3.3
packaging==20.4
Pillow==8.0.1
pip-review==1.1.0
PyJWT==2.0.0
pyparsing==2.4.7
pytz==2020.4
six==1.15.0
sqlparse==0.4.1
